Refactor the class under security pkgs
